apps/api/requirements.txt
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.3.3
pymysql==1.1.1
redis==5.1.1
oracledb==2.4.1
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.3
pytest-json-report==1.5.0
httpx==0.27.2
sqlglot==26.0.1
